+++ src/action_map.rs
@@ -15,29 +15,18 @@ use crate::{input_action::ActionOutput, prelude::*};
 ///
 /// Accessible from [`InputCondition::evaluate`] and [`InputModifier::apply`].
 #[derive(Default, Deref, DerefMut)]
-pub struct ActionMap(pub TypeIdMap<Action>);
+pub struct ActionMap(TypeIdMap<Action>);
 
 impl ActionMap {
     /// Returns associated state for action `A`.
     pub fn action<A: InputAction>(&self) -> Option<&Action> {
         self.get(&TypeId::of::<A>())
     }
-
-    /// Inserts a state for action `A`.
-    ///
-    /// Returns previously associated state if present.
-    pub fn insert_action<A: InputAction>(&mut self, action: Action) -> Option<Action> {
-        self.insert(TypeId::of::<A>(), action)
-    }
 }
 
 /// Data associated with an [`InputAction`] marker.
 ///
 /// Stored inside [`ActionMap`].
-///
-/// This struct could also be created manually to track state for an action
-/// with externally sourced data (e.g., network). Use [`Self::update`] to apply
-/// the data followed by [`Self::trigger_events`].
 #[derive(Clone, Copy)]
 pub struct Action {
     state: ActionState,
@@ -53,7 +42,7 @@ impl Action {
     ///
     /// [`Self::trigger_events`] will trigger events for `A`.
     #[must_use]
-    pub fn new<A: InputAction>() -> Self {
+    pub(crate) fn new<A: InputAction>() -> Self {
         Self {
             state: Default::default(),
             events: ActionEvents::empty(),
@@ -65,7 +54,7 @@ impl Action {
     }
 
     /// Updates internal state.
-    pub fn update(
+    pub(crate) fn update(
         &mut self,
         time: &Time<Virtual>,
         state: ActionState,
@@ -94,7 +83,7 @@ impl Action {
     /// Triggers events resulting from a state transition after [`Self::update`].
     ///
     /// See also [`Self::new`] and [`ActionEvents`].
-    pub fn trigger_events(&self, commands: &mut Commands, entity: Entity) {
+    pub(crate) fn trigger_events(&self, commands: &mut Commands, entity: Entity) {
         (self.trigger_events)(self, commands, entity);
     }